Add cargo, carrier and turret classes
